Update Backend/crit/src/main/resources/application.yml,...

See merge request s09-webmobile1-sub2/S09P12D201!72

+++ Backend/crit/src/main/java/com/ssafy/crit/boards/service/CommentService.java
@@ -1,38 +1,48 @@
 package com.ssafy.crit.boards.service;
 
-import com.ssafy.crit.boards.entity.Board;
-import com.ssafy.crit.boards.entity.Comment;
+import com.ssafy.crit.auth.entity.User;
+import com.ssafy.crit.auth.repository.UserRepository;
+import com.ssafy.crit.boards.entity.board.Board;
+import com.ssafy.crit.boards.entity.board.Comment;
 import com.ssafy.crit.boards.repository.BoardRepository;
 import com.ssafy.crit.boards.repository.CommentRepository;
-import com.ssafy.crit.boards.service.CommentDto;
-import com.ssafy.crit.imsimember.entity.Member;
+import com.ssafy.crit.boards.service.dto.CommentDto;
+import com.ssafy.crit.boards.service.dto.FileResponseDto;
+
+import com.ssafy.crit.common.error.code.ErrorCode;
+import com.ssafy.crit.common.error.exception.BadRequestException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
-
+/**
+ * author : 강민승
+ */
 @RequiredArgsConstructor
 @Service
 public class CommentService {
 
     private final CommentRepository commentRepository;
     private final BoardRepository boardRepository;
+    private final UserRepository userRepository;
 
     // 댓글 작성하기
     @Transactional
-    public CommentDto writeComment(Long boardId, CommentDto commentDto, Member member) {
-        Comment comment = new Comment();
-        comment.setContent(commentDto.getContent());
+    public CommentDto writeComment(Long boardId, CommentDto commentDto, User user) {
 
         // 게시판 번호로 게시글 찾기
         Board board = boardRepository.findById(boardId).orElseThrow(() -> {
-            return new IllegalArgumentException("게시판을 찾을 수 없습니다.");
+             return new BadRequestException(ErrorCode.NOT_EXISTS_BOARD_ID);
         });
 
-        comment.setMember(member);
-        comment.setBoard(board);
+        Comment comment = Comment.builder()
+                .content(commentDto.getContent())
+                .user(user)
+                .board(board)
+                .build();
+
         commentRepository.save(comment);
 
         return CommentDto.toDto(comment);
@@ -50,14 +60,20 @@ public class CommentService {
         return commentDtos;
     }
 
-
     // 댓글 삭제하기
     @Transactional
-    public String deleteComment(Long commentId) {
+    public String deleteComment(Long commentId, User user) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(()-> {
-            return new IllegalArgumentException("댓글 Id를 찾을 수 없습니다.");
+            return new BadRequestException(ErrorCode.NOT_EXISTS_BOARD_COMMENT);
         });
+
+        if(!comment.getUser().getId().equals(user.getId())){
+            throw new BadRequestException(ErrorCode.NOT_EXISTS_BOARD_AUTHORIZE);
+        }
+
         commentRepository.deleteById(commentId);
         return "삭제 완료";
     }
+
+
 }
